developer-hub-apis/pyproject.toml
[project]
name = "mcp-api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "com-precisely-apis>=18.1.1",
    "mcp[cli]>=1.9.2",
]
